Use forward slashes in paths on windows Move MPI fixtures into top level conftest.py Check for variation on language in example tests
