+++ tests/examples/__init__.py
@@ -79,10 +79,16 @@ class TestExample(base_class):
     def language(self, request, example_name, check_required_languages):
         r"""str: Language of the currect test."""
         avail_langs = get_example_languages(example_name)
-        if request.param not in avail_langs:
+        if request.param in avail_langs:
+            lang = request.param
+        elif request.param.lower() in avail_langs:
+            lang = request.param.lower()
+        elif request.param.upper() in avail_langs:
+            lang = request.param.upper()
+        else:
             pytest.skip(f"example dosn't have a {request.param} version")
         check_required_languages(
-            get_example_languages(example_name, language=request.param))
+            get_example_languages(example_name, language=lang))
         return request.param
 
     @pytest.fixture(scope="class", autouse=True)
