--- tests/examples/conftest.py
@@ -1,2 +1,0 @@
-# import MPI test fixtures
-from tests.communication.conftest import *  # noqa: F401, F403
